requirements.txt
flask>=3.0.0,<4
python-pptx==0.6.23
requests>=2.31,<3
Werkzeug